validate the actual chaincode type used by the blockchain
It turns out that the ndau chain doesn't use native chaincode
serialization for its chaincode sysvars. Possibly it wasn't
implemented when we started implementing the sysvar stuff?
Instead, chaincode values are stored and serialized as
wkt.Bytes instances: general purpose byte slices.

On the one hand, catching this sort of thing is exactly the point
of this PR. On the other hand, it sure would be nice to be able to
refactor that to use the native type; it's too bad that it appears
to be too late now.